Packetizing RTCM revealed a bug deep in the framework code -- one that only shows up when the device delivers less than one packet per cycle (e.g. normally 1 per second).  It became manifest because gpsfake, unable to packetize RTCM, inserts a delay between each *character* of an RTCM load.
The bug was this: the packet getter was returning nz only when it saw a 
complete packet.  What it needed to be doing (and does now) is return nz 
when it has new data, whether that new data finishes a packet or not.  
Otherwise the online check will spuriously fail if the device takes over a 
second to deliver a packet completion -- even if it's getting new data 
each time!


git-svn-id: svn://svn.berlios.de/gpsd/trunk@3002 3da76a2a-08e2-0310-9c3e-bc9091dd3f7d