General feeling seems to be: a) that checking for (newdata = read()) == -1 is good, b) that it is appropriate to retry in case of EINTR or EAGAIN c) a short delay in case of EINTR or EAGAIN is polite but not strictly needed.
I put the delay in because this may spin and I really hate busy waiting.
I may put in a retry counter: after some small number of retries that fail,
just give up and return 0 as if there was no data to be read. We shouldn't
have even reached this function unless select() said there was data to read
so it seems appropriate to get the data (with retries if necessary) unless a
truly fatal error occurred ((errno != EAGAIN) && (errno != EINTR))...

As we've just tagged a release, I feel better about committing this so this
gets a bit more testing.


git-svn-id: svn://svn.berlios.de/gpsd/trunk@3160 3da76a2a-08e2-0310-9c3e-bc9091dd3f7d