Move the checking of probe functions from within gpsd_open() out into its only caller, gpsd_activate().  This cleans up the layer separation between serial-device bashing and the driver-structure stuff.
The reeason for this change is that I want all the configurator and revert 
method calls close to each other in libgpsd_core.c so a reader can have
some hope of grokking the interactions correctly.  Petr Slansky thinks
there's a subtle bug in the current design, and I think he may be right.

Heads up, Gary!  Presently, the only driver this change could break is Garmin,
as you're the only user of the probe_detect hook.  I've changed gpsd_set_raw(),
of which you are also the only user, in what I believe is the correct way.
Test it, please...


git-svn-id: svn://svn.berlios.de/gpsd/trunk@3922 3da76a2a-08e2-0310-9c3e-bc9091dd3f7d