Pass fix data in url to the image generator, rather than making the image generator do its own query. This makes it easier to retry and wait for a receiver to come online if gpsd is running without '-n'.
Turn on buffering and do a couple of queries to make sure we have data to
display. More robust retry logic to follow.

Add simple hooks for plotting yourself with google maps. You will need to
enable this in the config file and get a google map api key.

(merge from my cvs)


git-svn-id: svn://svn.berlios.de/gpsd/trunk@3953 3da76a2a-08e2-0310-9c3e-bc9091dd3f7d