My change in r3187 unmasked a long-standing bug.  GPGGA and GPGLL need to set the mode field when they have valid fix data, otherwise the O reports will often show no fix on a GLL or GGA that arrives before GSA in the send cycle (and some devices, like the FV-18, don't normally send GSA at all).
This bug was being masked by an incorrect test; the have_fix()
function was returning positive if status had been set to nonzero and
no mode had yet been seen.  This, in turn, caused some misbehavior in
the P response; it would sometimes return P=? ?, which should not
actually have happened.  I spotted this symptom when looking at
reports from Chris Kuethe's PHP script.

This fix required rebuilding the regression tests, as the mode field on
GLL and GGA sentences now frequently has a different (higher) value than
it used to.

It is quite possible that this bug accounted for a lot of the 2D/3D
jitter that fans of the J patch were complaining about.

Screwing with the mode field like this in the absence of received GSAs is a
bit dodgy, but I don't see a better alternative.  NMEA 0183
brain-damage strikes again -- what a %@!$&# botch of a protocol!