This was dealt with in r3454.
I don't think that it's a bug on our part to have a certain
expectation about number formats - especially when we make that
explicit by forcing the use of the C locale on the wire.

It's not a bug in libc either - users have every right to ask the
computer to display data in a format they're comfortable with. In this
case, since the gpsd protocol is a wire protocol that happens to be
fairly human-readable too, the computer's preference is dominant. For
display engines (like cgps/xgps) the user preference can dominate.