1). I think the description of the co-variance matrix is wrong.
I expanded the squared terms so that s(n)^2 became s(n)*s(n)
and so on. I then swapped the first and second parts of some terms
so that the columns formed a regular pattern. I then "fixed" the
terms which seemed wrong and simplifed the s(n)*s(n) terms back
to s(n)^2.

This does not affect the code, which implemented the correct
matrix anyway.

2)
The reporting of DOPS had the text descrtiption of calculated
and reported fields reversed.

3)
The reporting of DOPS had a duplcated pair of format fields
for the TDOP. The format now matches the data.


git-svn-id: svn://svn.berlios.de/gpsd/trunk@4541 3da76a2a-08e2-0310-9c3e-bc9091dd3f7d