Bugs can lurk indefinitely in the damnedest places.  This patch fixes a bonehead error in gpsd's input data management that has been undetected for years because it only showed up reproducibly when one of gpsd's device inputs was connected to a file, *not* an actual GPS constantly streaming data to it.
The error was that packet_get() would return 0, signaling end of input
in some circumstances, any time a physical read from the device failed
or returned 0 - even if there was still data in the packet input buffer.

As a result, gpsd could fail to see a random number of bytes at the
end of a file source.  This probably accounts for sporadic spurious
errors associated with short reads in the regression tests. The error
was masked when connected to actual GPSes, which would always make
more data available for physical read within a second or so.

The new test tries to parse out packet data if either the physical read 
returned > 0 or the packet input buffer is nonempty.  Note, this depends 
on the packet_parse() function shortening the input buffer by at least one
byte per call every time, otherwise this test could result in gpsd 
infinite-looping on an empty buffer.


git-svn-id: svn://svn.berlios.de/gpsd/trunk@4766 3da76a2a-08e2-0310-9c3e-bc9091dd3f7d