I have no idea what crack-addled rhesus monkey wrote this crap but it's wrong.
Oh wait, that was me.

Not sure what I was thinking but "0xffff ^ getub(buf, 7 + 16)" makes no sense.
It's the wrong place to look for the number of PRNs used, and the wrong size
too - none of the elements in the message are bytes, they're 16-bit or 32-bit.

I think this may be where my bus error in dop() was coming from. According to
gdb, somehow satellites_used was in the 65k range.


git-svn-id: svn://svn.berlios.de/gpsd/trunk@5384 3da76a2a-08e2-0310-9c3e-bc9091dd3f7d