gps.py: Make poll() read all buffered data before returning
This might well be the most controversial change. I have rewritten much
of the reading code in order to make sure that a client receives all
data sent by GPSd. I was using a client which slept for a few seconds
between calls to gpsdata.poll(), which had two effects:

1. The client was more and more behind with reading messages
   because it only ever read the next line from GPSd. The most
   obvious thing was that the GPS time displayed advanced only one
   second per sleep interval.

2. The TCP input queue grew non-stop because data sent by GPSd
   was not read fast enough.

With this change, readlines() and readstanzas() become generator
functions that return whatever has not yet been read by the client.
Every time poll() is called, it reads the whole input queue until
it's empty and updates the GPS data from each received message.
So a client can poll GPSd as infrequently as it wants and still
gets everything.

git-svn-id: svn://svn.berlios.de/gpsd/trunk@5584 3da76a2a-08e2-0310-9c3e-bc9091dd3f7d